Added files via upload from laptop.
It contains a number of Java files.
NameNode - A file that has a list of constructors for the NameNode attributes.
DataNode - A file that has a list of constructors for the DataNode attributes.
DataModel - A file that has a list of constructors for the structure of a file. A list of attributes and one of them being a List<DataBlocks>
DataBlock - A file that has a list of constructors for the DataBlock attributes.
DataParser - The file used to parse the raw data that we will be feeding it.

HDFS_FSCK_BLOCK_LOCATIONS - This is the raw dummy data from a Hadoop server last year.